+++ demos/unpaired_ct_abdomen/demo_predict.py
@@ -1,30 +1,54 @@
-"""
-This is a palceholder python file for calling the installed functions
-"""
+import argparse
+from datetime import datetime
 
-print(
-    "Please use command line functions for this demo \
-    by following the instructions in readme.md"
-)
-
-"""
 from deepreg.predict import predict
 
-log_dir = "unpaired_ct_abdomen_log_comb"
-ckpt_path = "demos/unpaired_ct_abdomen/dataset/pre-trained/weights-epoch100.ckpt"
-gpu = ""
-gpu_allow_growth = False
+name = "unpaired_ct_abdomen"
+ckpt_index_dict = {"comb": 2000, "unsup": 5000, "weakly": 2250}
 
-config_path = "demos/unpaired_ct_abdomen/unpaired_ct_abdomen_comb.yaml"
+parser = argparse.ArgumentParser()
+parser.add_argument(
+    "--method",
+    help="Training method, comb or unsup or weakly",
+    type=str,
+    required=True,
+)
+args = parser.parse_args()
+method = args.method
+assert method in [
+    "comb",
+    "unsup",
+    "weakly",
+], f"method should be comb or unsup or weakly, got {method}"
+
+ckpt_index = ckpt_index_dict[method]
+print(
+    "\n\n\n\n\n"
+    "=========================================================\n"
+    "The prediction can also be launched using the following command.\n"
+    "deepreg_predict --gpu '' "
+    f"--config_path demos/{name}/{name}_{method}.yaml "
+    f"--ckpt_path demos/{name}/dataset/pretrained/{method}/weights-epoch{ckpt_index}.ckpt "
+    f"--log_root demos/{name} "
+    f"--log_dir logs_predict/{method} "
+    "--save_png --mode test\n"
+    "=========================================================\n"
+    "\n\n\n\n\n"
+)
+
+log_root = f"demos/{name}"
+log_dir = f"logs_predict/{method}/" + datetime.now().strftime("%Y%m%d-%H%M%S")
+ckpt_path = f"{log_root}/dataset/pretrained/{method}/weights-epoch{ckpt_index}.ckpt"
+config_path = f"{log_root}/{name}_{method}.yaml"
 
 predict(
-    gpu=gpu,
-    gpu_allow_growth=gpu_allow_growth,
+    gpu="0",
+    gpu_allow_growth=False,
     ckpt_path=ckpt_path,
     mode="test",
     batch_size=1,
+    log_root=log_root,
     log_dir=log_dir,
     sample_label="all",
     config_path=config_path,
 )
-"""